Pull reusable code from errors.rs into cargo_errors.rs
I intent to extract this into another library for rustup's use.
